Implement HP bar Allow TowerSeletor to capture various values of enemy

+++ src/Tower/ArcherTower.java
@@ -6,8 +6,8 @@ public class ArcherTower implements Tower{
     double cost;
     double damage;
     int level;
-    double alertRange;
-    public ArcherTower(double cost, double damage, int level, double alertRange) {
+    int alertRange;
+    public ArcherTower(double cost, double damage, int level, int alertRange) {
         this.cost = cost;
         this.damage = damage;
         this.level = level;
@@ -26,7 +26,7 @@ public class ArcherTower implements Tower{
         return level;
     }
     @Override
-    public double getAlertRange() {
+    public int getAlertRange() {
         return alertRange;
     }
     public void update(ConcreteEnemy enemy, String status) {
